+++ CHANGELOG.md
@@ -13,6 +13,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Improve performance of login with hardware wallets
 
+### Fixed
+- Create 2of3 subaccount with ledger
+
 ## [3.8.4] - 2022-06-17
 
 ### Added