+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 - Create 2of3 subaccount with ledger
+- Fix hardware wallets logout view hierarchy
 
 ## [3.8.4] - 2022-06-17